ocr-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytesseract>=0.3.8
Pillow>=8.3.2
numpy>=1.21.0
opencv-python-headless>=4.5.3
